Remove config from clients
